refactored code, deleted unused portions, added applicable comments
